+++ packages/helpers/src/index.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { isFunction, isObject } from 'lodash';
+import { isFunction, isObject, isPlainObject } from 'lodash';
 
 import {
   AsFieldContext,
@@ -10,7 +10,7 @@ import {
   withSlide,
   withFormSlideField,
 } from './contexts';
-import { default as required, createRequired } from './required';
+import required, { createRequired } from './required';
 import createFormatter from './createFormatter';
 import keyCodes from './keyCodes';
 
@@ -38,11 +38,13 @@ export type Omit<T, K> = Pick<T, Exclude<keyof T, K>>;
  */
 export type Subtract<T, K> = Omit<T, keyof K>;
 
+/* eslint-disable no-nested-ternary */
 /**
  * Executes a the first arg as a function if it is a function with supplied arguments
  * Returns the first arg if not a function
  */
 export const execIfFunc = (fn: any, ...args: any[]) => (isFunction(fn) ? (args.length > 0 ? fn(...args) : fn()) : fn);
+/* eslint-enable no-nested-ternary */
 
 /**
  * Executes the first arg if a function with remaining args; else, returns an array of remaining args
@@ -58,25 +60,25 @@ export const execOrMapFn = (fn: any | any[], ...args: any[]) =>
 /**
  * Provide this to be classNames
  */
-export function classes(...args: any[]) {
-  const arr = args.reduce((c: any[], obj) => {
-    if (Array.isArray(obj)) {
-      return [...c, ...obj.filter(Boolean)];
-    }
-    if (isObject(obj)) {
-      return [
-        ...c,
-        ...Object.keys(obj)
-          // @ts-ignore: this is an object
-          .map(k => Boolean(obj[k]) && k)
-          .filter(Boolean),
-      ];
-    }
-
-    return [...c, typeof obj === 'string' ? obj : ''];
-  }, []);
-  return arr.filter(Boolean).join(' ');
-}
+export const classes = (...args: any[]): string =>
+  args
+    .reduce((c: any[], obj) => {
+      if (Array.isArray(obj)) {
+        return [...c, ...obj.filter(Boolean)];
+      }
+      if (isPlainObject(obj)) {
+        return [
+          ...c,
+          ...Object.keys(obj)
+            // @ts-ignore: this is an object
+            .map(k => Boolean(obj[k]) && k)
+            .filter(Boolean),
+        ];
+      }
+      return [...c, typeof obj === 'string' ? obj : ''];
+    }, [])
+    .filter(Boolean)
+    .join(' ');
 
 /**
  * Checks if the argument is defined
@@ -92,7 +94,7 @@ export const isNull = (arg: any) => arg === null;
  * Checks if the argument is `thennable`
  */
 export const isPromise = (obj: any): boolean =>
-  !!obj && ['function', 'object'].includes(typeof obj) && isFunction(obj.then);
+  obj && ['function', 'object'].includes(typeof obj) && isFunction(obj.then);
 
 /**
  * Wraps an object in a Promise.resolve if it is not thennable
@@ -167,6 +169,12 @@ export const gatherErrors = (
     .filter(Boolean);
 
 /**
+ * Ensures that a `val` is an array filtered of falsy values
+ */
+export const alwaysFilteredArray = <P extends any>(val: any | any[]): P[] =>
+  (Array.isArray(val) ? val : [val]).filter(Boolean);
+
+/**
  * Attempts to transform a ReactNode into a uniqueish key
  */
 export const toKey = (arg: React.ReactNode): string | number => {
@@ -207,6 +215,7 @@ export function moveCursor(el: any, position = -1) {
   // @ts-ignore
   if (canAccessSelectionStart(el.type)) {
     const pos = position > -1 ? position : el.value.length;
+    // eslint-disable-next-line no-multi-assign
     el.selectionStart = el.selectionEnd = pos;
   } else if ('createTextRange' in el && isFunction(el.createTextRange)) {
     const range = el.createTextRange();