+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
-## master
+## 0.3.2
 
+- [all] update dependencies
+- [all] update lodash imports to allow rollup to treeshake better
 - [slider] _breaking_ change `next` behavior so that it calls the submit instead of going to the last slide when no viable candidates are found
 - [field] allow for `asField` to set a default value for a field type
 - [field] update file structures to appease the linter
+- [helpers] slim down classes code
 
 ## 0.3.1